refactor: 💡 modified field option merge module
This commit will update field option module
